+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- You can now declare multiple unconditional styles by passing an array of stylesets/class names
 - You can now add extra styles to a ZACS component via `zacs:style={{ attr: value }}` syntax
 - Strip ZACS declarations (`const Foo = zacs.foo(...)`) from output code by default. Pass `keepDeclarations: true` in Babel config to revert to previous behavior